# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project WebAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/activity": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "List of activity",
                "description": "List of activity",
                "operationId": "list_activity",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/activity/{uuid}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get activity",
                "description": "Get activity",
                "operationId": "get_activity",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "activity uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Account"
                ],
                "summary": "Delete activity",
                "description": "Delete activity",
                "operationId": "delete_activity",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "activity uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/activity/user/{uuid}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "List of user activity",
                "description": "List of user activity",
                "operationId": "list_user_activity",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "user uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/activity/entity/{uuid}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "List of entity activity",
                "description": "List of entity activity",
                "operationId": "list_entity_activity",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "entity uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/department": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "List of department",
                "description": "List of department",
                "operationId": "list_department",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/department/{uuid}": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Get department",
                "description": "Get department",
                "operationId": "get_department",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "department uuid",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/hosting": {
            "get": {
                "tags": [
                    "Helper"
                ],
                "summary": "List of hosting",
                "description": "List of hosting",
                "operationId": "list_hosting",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/role": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "List of role",
                "description": "List of role",
                "operationId": "list_role",
                "responses": {
                    "200": {
                        "description": "Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}